pyproject.toml
[project]
name = "myapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.116.2",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "pillow>=11.3.0",
    "uvicorn>=0.35.0",
]
